+++ src/pages/create/addBillStep/styles/BillList.styles.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+import { VStack } from '@chakra-ui/react';
+
+export const BillList = styled(VStack)`
+  width: 100%;
+  padding: 1rem 1.25rem;
+`;
